Defer call to onLoadFinished for Loader when cursor data is identical
