+++ app/build.gradle
@@ -137,7 +137,7 @@ dependencies {
 
     implementation 'eu.chainfire:libsuperuser:1.0.0.+'
 
-    implementation 'com.cloudrail:cloudrail-si-android:2.22.1'
+    implementation 'com.cloudrail:cloudrail-si-android:2.22.4'
 
     implementation 'com.github.PhilJay:MPAndroidChart:v3.0.2'//Nice charts and graphs
